extract methods to new parent class AbstractBuilder
